Rename download.py to more fitting upload.py
